+++ src/library/DIPS.Mobile.UI/API/Accessibility/ControlType.cs
@@ -5,5 +5,6 @@ public enum ControlType
     None = 0,
     Button = 1,
     Label = 2,
-    Choice = 3
+    Choice = 3,
+    Input = 4
 }